requirements.txt
pandas>=1.5.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
gradio>=4.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
requests>=2.28.0
